+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 22.7.0
+
+* [swift, kotlin] Adds event channel support.
+* [swift, kotlin] Adds `sealed` class inheritance support.
+* [swift] Updates codec class names to be upper camel case.
+
 ## 22.6.4
 
 * [swift] Fixes the channel names of the named constructors of ProxyApis.